trying move to plain

+++ suckless/slstatus/components/disk.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <sys/statvfs.h>
 
+#include "../slstatus.h"
 #include "../util.h"
 
 const char *
@@ -28,7 +29,7 @@ disk_perc(const char *path)
 	}
 
 	return bprintf("%d", (int)(100 *
-	               (1.0f - ((float)fs.f_bavail / (float)fs.f_blocks))));
+	               (1 - ((double)fs.f_bavail / (double)fs.f_blocks))));
 }
 
 const char *
